% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_dups}
\alias{filter_dups}
\title{Return dataframe containing duplicate values}
\usage{
filter_dups(df, var, n=1)
}
\arguments{
\item{df}{Dataframe}

\item{var}{The column name for which to filter the duplicates}

\item{n}{Return rows containing more then n instances of the specified column.}
}
\value{
A dataframe containing the rows filtered for duplicates in the specified
column.
}
\description{
Uses tidyeval. This function takes a dataframe and a column in the dataframe, and
returns the dataframe filtered for duplicate values in the specified column.
}
