% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{gglegend_size}
\alias{gglegend_size}
\title{Convinience function for modifying ggplot legend sizes in a consistent way}
\usage{
gglegend_size(size)
}
\arguments{
\item{size}{Size of the title, text, and boxes of the legend, in pt}
}
\description{
Convinience function for modifying ggplot legend sizes in a consistent way
}
