% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figtex.R
\name{make_plot_file}
\alias{make_plot_file}
\title{Create file name and print plot}
\usage{
make_plot_file(type, plot, figure.dir, title, res, width, height)
}
\description{
Internal function for \code{save_figtex}
}
