% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quick_install}
\alias{quick_install}
\title{Install the Rfunctions package}
\usage{
quick_install()
}
\description{
Shortcut to document and install the R package at 'bin/Rfunction'.
Also, reload Rstudio.
}
