% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twentycol}
\alias{twentycol}
\title{Return named vector of 20 custom colors}
\usage{
twentycol()
}
\value{
Named list of 20 colors
}
\description{
Use this function to return a custom named list of 20 colors. These are somewhat
useful for metagenomic graphs.
}
