% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figtex.R
\name{write_latex_string}
\alias{write_latex_string}
\title{Write latex string for figure generation}
\usage{
write_latex_string(caption, figure.path, title)
}
\arguments{
\item{caption}{The caption}

\item{figure.path}{The relative path of the figure to be included}

\item{title}{The title of the figure.}
}
\description{
Internal function, write the string for the tex input file for a
generated figure
}
